+++ app/components/accounts/portfolios_component.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Accounts
+  class PortfoliosComponent < ViewComponent::Base
+    def initialize(portfolios:)
+      @portfolios = portfolios
+    end
+  end
+end