+++ app/controllers/account_controller.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class AccountController < ApplicationController
+  before_action :portfolios
+
+  def index; end
+
+  private
+
+  def portfolios
+    @portfolios = Portfolios::Fetching::ForAccountService.call(user: current_user).result
+  end
+end