+++ CHANGELOG.md
@@ -16,3 +16,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - new frontend structure with controllers, components
 - SearchSphinx as search engine
 - selling securities
+- creating portfolios on account page